+++ packages/spacecat-shared-data-access/test/it/configuration/configuration.test.js
@@ -25,7 +25,8 @@ describe('Configuration IT', async () => {
   let sampleData;
   let Configuration;
 
-  before(async () => {
+  before(async function () {
+    this.timeout(10000);
     sampleData = await seedDatabase();
 
     const dataAccess = getDataAccess();